Major changes to voronoi, restructuring

+++ AIProject/AIProject/_config.h
@@ -33,9 +33,9 @@ enum DifficultyMode
 	// Radius Computation Buffer
 	const int RADIUS_BUFFER = 10;
 	// Node Heuristic Weightage: Distance from <Home Agent> to <Path Node>
-	const float NODE_WEIGHTAGE_D1 = 0.3;
+	const double NODE_WEIGHTAGE_D1 = 0.3;
 	// Node Heuristic Weightage: Distance from <Path Node> to <Goal>
-	const float NODE_WEIGHTAGE_D2 = 0.7;
+	const double NODE_WEIGHTAGE_D2 = 0.7;
